requirements.txt
django >=1.11, <2
django-environ==0.4.4
whitenoise==3.3.1
psycopg2>=2.7.4, <2.8.0
django-cron==0.5.0
ping3==1.0.0
django-solo==1.1.3
numpy>=1.9.0
pandas>=0.22.0
